requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0
openai==0.27.0
aiohttp==3.8.1
python-json-logger==2.0.2
tavily-python==0.1.7
requests==2.26.0